rename dosbox to OPL
