+++ sdl2-sonic-drivers/src/hardware/opl/Config.cpp
@@ -2,7 +2,7 @@
 #include <cstring>
 #include <spdlog/spdlog.h>
 #include <hardware/opl/scummvm/mame/OPL.hpp>
-#include <hardware/opl/scummvm/dosbox/dosbox.hpp>
+#include <hardware/opl/scummvm/dosbox/OPL.hpp>
 #include <hardware/opl/scummvm/nuked/OPL.hpp>
 #include <hardware/opl/woody/WoodyOPL.hpp>
 #include <hardware/opl/mame/MameOPL.hpp>
